+++ eras/conway/impl/CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## 1.19.0.0
 
+* Add `MemPack` instance for `PlutusScript ConwayEra`
 * Deprecate `Conway` type synonym
 * Remove crypto parametrization from `ConwayEra`
+
+### `testlib`
+
+* Add `Arbitrary` instance for `ConwayBbodyPredFailure` and `ConwayMempoolPredFailure`
 
 ## 1.18.0.0